hashbusteravalon: Klondike-based metadriver
